blockstorage: Merge volumeactions extension
This builds upon the '/volumes' resource so the code is forklifted into
the 'openstack/volume/{v2,v3}/volumes' module.
