+++ internal/acceptance/openstack/blockstorage/v2/volumes_test.go
@@ -9,8 +9,8 @@ import (
 	"time"
 
 	"github.com/gophercloud/gophercloud/v2/internal/acceptance/clients"
+	compute "github.com/gophercloud/gophercloud/v2/internal/acceptance/openstack/compute/v2"
 	"github.com/gophercloud/gophercloud/v2/internal/acceptance/tools"
-	"github.com/gophercloud/gophercloud/v2/openstack/blockstorage/extensions/volumeactions"
 	"github.com/gophercloud/gophercloud/v2/openstack/blockstorage/v2/snapshots"
 	"github.com/gophercloud/gophercloud/v2/openstack/blockstorage/v2/volumes"
 	th "github.com/gophercloud/gophercloud/v2/testhelper"
@@ -76,7 +76,7 @@ func TestVolumesCreateForceDestroy(t *testing.T) {
 
 	tools.PrintResource(t, newVolume)
 
-	err = volumeactions.ForceDelete(context.TODO(), client, newVolume.ID).ExtractErr()
+	err = volumes.ForceDelete(context.TODO(), client, newVolume.ID).ExtractErr()
 	th.AssertNoErr(t, err)
 }
 
@@ -132,5 +132,182 @@ func TestVolumesCascadeDelete(t *testing.T) {
 	th.AssertNoErr(t, err)
 
 	t.Logf("Successfully deleted volume: %s", vol.ID)
-
 }
+
+func TestVolumeActionsUploadImageDestroy(t *testing.T) {
+	clients.SkipReleasesAbove(t, "stable/ocata")
+
+	blockClient, err := clients.NewBlockStorageV2Client()
+	th.AssertNoErr(t, err)
+
+	computeClient, err := clients.NewComputeV2Client()
+	th.AssertNoErr(t, err)
+
+	volume, err := CreateVolume(t, blockClient)
+	th.AssertNoErr(t, err)
+	defer DeleteVolume(t, blockClient, volume)
+
+	volumeImage, err := CreateUploadImage(t, blockClient, volume)
+	th.AssertNoErr(t, err)
+
+	tools.PrintResource(t, volumeImage)
+
+	err = DeleteUploadedImage(t, computeClient, volumeImage.ImageID)
+	th.AssertNoErr(t, err)
+}
+
+func TestVolumeActionsAttachCreateDestroy(t *testing.T) {
+	clients.SkipReleasesAbove(t, "stable/ocata")
+
+	blockClient, err := clients.NewBlockStorageV2Client()
+	th.AssertNoErr(t, err)
+
+	computeClient, err := clients.NewComputeV2Client()
+	th.AssertNoErr(t, err)
+
+	server, err := compute.CreateServer(t, computeClient)
+	th.AssertNoErr(t, err)
+	defer compute.DeleteServer(t, computeClient, server)
+
+	volume, err := CreateVolume(t, blockClient)
+	th.AssertNoErr(t, err)
+	defer DeleteVolume(t, blockClient, volume)
+
+	err = CreateVolumeAttach(t, blockClient, volume, server)
+	th.AssertNoErr(t, err)
+
+	newVolume, err := volumes.Get(context.TODO(), blockClient, volume.ID).Extract()
+	th.AssertNoErr(t, err)
+
+	DeleteVolumeAttach(t, blockClient, newVolume)
+}
+
+func TestVolumeActionsReserveUnreserve(t *testing.T) {
+	clients.SkipReleasesAbove(t, "stable/ocata")
+
+	client, err := clients.NewBlockStorageV2Client()
+	th.AssertNoErr(t, err)
+
+	volume, err := CreateVolume(t, client)
+	th.AssertNoErr(t, err)
+	defer DeleteVolume(t, client, volume)
+
+	err = CreateVolumeReserve(t, client, volume)
+	th.AssertNoErr(t, err)
+	defer DeleteVolumeReserve(t, client, volume)
+}
+
+func TestVolumeActionsExtendSize(t *testing.T) {
+	clients.SkipReleasesAbove(t, "stable/ocata")
+
+	blockClient, err := clients.NewBlockStorageV2Client()
+	th.AssertNoErr(t, err)
+
+	volume, err := CreateVolume(t, blockClient)
+	th.AssertNoErr(t, err)
+	defer DeleteVolume(t, blockClient, volume)
+
+	tools.PrintResource(t, volume)
+
+	err = ExtendVolumeSize(t, blockClient, volume)
+	th.AssertNoErr(t, err)
+
+	newVolume, err := volumes.Get(context.TODO(), blockClient, volume.ID).Extract()
+	th.AssertNoErr(t, err)
+
+	tools.PrintResource(t, newVolume)
+}
+
+func TestVolumeActionsImageMetadata(t *testing.T) {
+	clients.SkipReleasesAbove(t, "stable/ocata")
+
+	blockClient, err := clients.NewBlockStorageV2Client()
+	th.AssertNoErr(t, err)
+
+	volume, err := CreateVolume(t, blockClient)
+	th.AssertNoErr(t, err)
+	defer DeleteVolume(t, blockClient, volume)
+
+	err = SetImageMetadata(t, blockClient, volume)
+	th.AssertNoErr(t, err)
+}
+
+func TestVolumeActionsSetBootable(t *testing.T) {
+	clients.SkipReleasesAbove(t, "stable/ocata")
+
+	blockClient, err := clients.NewBlockStorageV2Client()
+	th.AssertNoErr(t, err)
+
+	volume, err := CreateVolume(t, blockClient)
+	th.AssertNoErr(t, err)
+	defer DeleteVolume(t, blockClient, volume)
+
+	err = SetBootable(t, blockClient, volume)
+	th.AssertNoErr(t, err)
+}
+
+func TestVolumeActionsResetStatus(t *testing.T) {
+	clients.SkipReleasesAbove(t, "stable/ocata")
+
+	client, err := clients.NewBlockStorageV2Client()
+	th.AssertNoErr(t, err)
+
+	volume, err := CreateVolume(t, client)
+	th.AssertNoErr(t, err)
+	defer DeleteVolume(t, client, volume)
+
+	tools.PrintResource(t, volume)
+
+	err = ResetVolumeStatus(t, client, volume, "error")
+	th.AssertNoErr(t, err)
+
+	err = ResetVolumeStatus(t, client, volume, "available")
+	th.AssertNoErr(t, err)
+}
+
+// Note(jtopjian): I plan to work on this at some point, but it requires
+// setting up a server with iscsi utils.
+/*
+func TestVolumeConns(t *testing.T) {
+    clients.SkipReleasesAbove(t, "stable/ocata")
+
+    client, err := newClient()
+    th.AssertNoErr(t, err)
+
+    t.Logf("Creating volume")
+    cv, err := volumes.Create(client, &volumes.CreateOpts{
+        Size: 1,
+        Name: "blockv2-volume",
+    }).Extract()
+    th.AssertNoErr(t, err)
+
+    defer func() {
+        err = volumes.WaitForStatus(client, cv.ID, "available", 60)
+        th.AssertNoErr(t, err)
+
+        t.Logf("Deleting volume")
+        err = volumes.Delete(client, cv.ID, volumes.DeleteOpts{}).ExtractErr()
+        th.AssertNoErr(t, err)
+    }()
+
+    err = volumes.WaitForStatus(client, cv.ID, "available", 60)
+    th.AssertNoErr(t, err)
+
+    connOpts := &volumes.ConnectorOpts{
+        IP:        "127.0.0.1",
+        Host:      "stack",
+        Initiator: "iqn.1994-05.com.redhat:17cf566367d2",
+        Multipath: false,
+        Platform:  "x86_64",
+        OSType:    "linux2",
+    }
+
+    t.Logf("Initializing connection")
+    _, err = volumes.InitializeConnection(client, cv.ID, connOpts).Extract()
+    th.AssertNoErr(t, err)
+
+    t.Logf("Terminating connection")
+    err = volumes.TerminateConnection(client, cv.ID, connOpts).ExtractErr()
+    th.AssertNoErr(t, err)
+}
+*/
