--- openstack/blockstorage/extensions/volumeactions/testing/doc.go
@@ -1,2 +1,0 @@
-// volumeactions unit tests
-package testing
