+++ openstack/blockstorage/v2/volumes/testing/doc.go
@@ -1,2 +1,2 @@
-// volumes_v2
+// volumes unit tests
 package testing
