+++ openstack/blockstorage/v3/volumes/testing/doc.go
@@ -1,2 +1,2 @@
-// volumes_v3
+// volumes unit tests
 package testing
